Lesson_5_Task_1.1

+++ Lesson_5/Model.java
@@ -1,8 +1,8 @@
 package Lesson_5;
 
 public interface Model {
-    int result(char act);
-    void setX(int val);
-    void setY(int val);
+    double result(char act);
+    void setX(double val);
+    void setY(double val);
 
 }
